pyproject.toml
[tool.poetry]
name = "vello"
version = "0.1.0"
description = ""
authors = ["Igwe Daniel <danielchibuzoigwe@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
Django = "^5.1.2"
djangorestframework = "^3.15.2"
django-filter = "^24.3"
djoser = "^2.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
